feat: adding benchmarks for proto serialization
results from bench tests:

	test perfetto::tests::bench_create_example_trace ... bench:   2,149,230 ns/iter (+/- 722,057)
	test tests::bench_add_example_profile            ... bench:     936,466 ns/iter (+/- 179,738)

without deserialization, protos take ~ 1/2 of the time.
without serialization or deserialization, protos take ~/10 of the time.

So it's clear that flatbuffers work best when it's expected for both a
serialization and a deserialization to occur in at the same time. Or,
incurring a slight overhead on serialization for a much lower overhead
on deserialization later.